+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+v0.03:
+* gbs - fix get tag duration fading if "," in title
+* gbs - fix false stereo detection
+* support of NES files = nfs (with or without m3u)
+
 v0.02:
 * improve tags loops, especialy for artist
 * support of PS1 files = pfs, minipfs
@@ -14,6 +19,6 @@ v0.01:
 * support of PS1 files in = xa
 * support of Atari ST files in = snd|sndh
 * support of RAW PCM files in = bin|pcm|raw|tak
-* support of game boy files in = gbs
+* support of game boy files in = gbs (with or without m3u)
 * support of various console files in = s98|vgm|vgz
 * support of various console files in = aa3|adp|adpcm|ads|adx|aif|aifc|aix|ast|at3|bcstm|bcwav|bfstm|bfwav|cfn|dsp|eam|fsb|genh|his|hps|imc|int|laac|ktss|msf|mtaf|mib|mus|rak|raw|sad|sfd|sgd|sng|spsd|str|ss2|thp|vag|vgs|vpk|wem|xvag|xwav